create a folder module
