+++ modules/tasks.js
@@ -1,0 +1,24 @@
+const tasks = [
+  {
+    index: 1,
+    description: 'complete part1 todo list project',
+    completed: false,
+  },
+  {
+    index: 2,
+    description: 'complete part2 todo list project',
+    completed: false,
+  },
+  {
+    index: 3,
+    description: 'share my project ToDolist on twitter',
+    completed: true,
+  },
+  {
+    index: 4,
+    description: 'add two function add and remove',
+    completed: false,
+  },
+];
+
+export default tasks;
