+++ src/index.js
@@ -1,27 +1,5 @@
 import './style.css';
-
-const tasks = [
-  {
-    index: 1,
-    description: 'complete part1 todo list project',
-    completed: false,
-  },
-  {
-    index: 2,
-    description: 'complete part2 todo list project',
-    completed: false,
-  },
-  {
-    index: 3,
-    description: 'share my project ToDolist on twitter',
-    completed: true,
-  },
-  {
-    index: 4,
-    description: 'add two function add and remove',
-    completed: false,
-  },
-];
+import tasks from '../modules/tasks.js';
 
 const getTaskItemList = () => {
   let taskItemList = '';
